Introduced new api hooks file
